requirements.txt
aiogram
supabase==2.15.3
python-dotenv
httpx
matplotlib
pandas
